+++ taxi_project/myAPP/utils.py
@@ -1,0 +1,32 @@
+from pathlib import Path
+from django.core.exceptions import ImproperlyConfigured
+
+import requests
+import json
+
+# secret.json에서 카카오 키 가져오기
+BASE_DIR = Path(__file__).resolve().parent.parent
+
+secret_file = BASE_DIR / 'secrets.json'
+
+with open(secret_file) as file:
+    secrets = json.loads(file.read())
+
+def get_secret(setting,secrets_dict = secrets):
+    try:
+        return secrets_dict[setting]
+    except KeyError:
+        error_msg = f'Set the {setting} environment variable'
+        raise ImproperlyConfigured(error_msg)
+
+KAKAO_KEY = get_secret('KAKAO_KEY') 
+
+def coordinate_send_request(road_address):
+    url = 'https://dapi.kakao.com/v2/local/search/address?query={address}'.format(address=road_address)  # 요청을 보낼 API의 URL
+    headers = {
+        'Authorization': 'KakaoAK {}'.format(KAKAO_KEY),  # Authorization 헤더
+        'Content-Type': 'application/json'  # 필요에 따라 Content-Type 헤더 추가
+    }
+
+    result = json.loads(str(requests.get(url, headers=headers).text))
+    return result
