+++ taxi_project/myAPP/views.py
@@ -12,30 +12,9 @@ from rest_framework_simplejwt.tokens import RefreshToken
 from drf_yasg.utils import swagger_auto_schema
 from drf_yasg import openapi
 
-from pathlib import Path
-from django.core.exceptions import ImproperlyConfigured
-
-import requests
-import json
 from .models import *
 from .serializer import *
-
-# secret.json에서 카카오 키 가져오기
-BASE_DIR = Path(__file__).resolve().parent.parent
-
-secret_file = BASE_DIR / 'secrets.json'
-
-with open(secret_file) as file:
-    secrets = json.loads(file.read())
-
-def get_secret(setting,secrets_dict = secrets):
-    try:
-        return secrets_dict[setting]
-    except KeyError:
-        error_msg = f'Set the {setting} environment variable'
-        raise ImproperlyConfigured(error_msg)
-
-KAKAO_KEY = get_secret('KAKAO_KEY') 
+from .utils import coordinate_send_request
 
 @swagger_auto_schema(
         method="POST", 
@@ -111,19 +90,12 @@ def new_taxi(request):
 @api_view(['POST'])
 @permission_classes([AllowAny])
 def coordinate(request):
-    #road_address = request.data.get('road_address')
     serializer = RoadAddressSerializer(data = request.data)
     # api 호출
 
     if serializer.is_valid():
         road_address = serializer.data['road_address']
-
-        url = 'https://dapi.kakao.com/v2/local/search/address?query={address}'.format(address=road_address)  # 요청을 보낼 API의 URL
-        headers = {
-            'Authorization': 'KakaoAK {}'.format(KAKAO_KEY),  # Authorization 헤더
-            'Content-Type': 'application/json'  # 필요에 따라 Content-Type 헤더 추가
-        }
-
-        result = json.loads(str(requests.get(url, headers=headers).text))
+        result = coordinate_send_request(road_address)
         return Response({"road_address":road_address, "latitude":result["documents"][0]['y'], "longitude":result["documents"][0]['x']}, status=201)
     return Response({'status':'400','message':serializer.errors}, status=400)
+
